Camera + Map
Spent quite a while reworking the camera to go from a static camera to one which has look ahead based on your mouse position and that will slow down when making turns or else to give emphasis to the action. Also added a crude pixel art map and play zone (red). Needs to be reorganized, but should work for now.